move fitmethod into fitting_utils(to be moved)
